requirements.txt
flet==0.23.2
configparser==6.0.0
google-generativeai==0.3.2
obsws-python==1.6.2
openai
pytchat==0.5.5
python-dotenv==1.0.0
requests==2.31.0
sounddevice==0.4.6
soundfile==0.12.1
tiktoken==0.6.0
websocket-client==1.7.0
numpy==1.26.3
pandas==2.2.1
python-dotenv